requirements.txt
Django==2.2.16
djangorestframework==3.12.4
django-extra-fields==3.0.2
django-filter==2.4.0
djoser==2.1.0
isort==5.10.1
pytz==2022.6
Pillow==9.3.0
python-dotenv==0.20.0
psycopg2-binary==2.8.6
sqlparse==0.4.3